Add {write_concurrency, auto} option to ETS tables
This commit makes it possible for users to configure ETS tables with
the {write_concurrency, auto} option. This option forces tables to
automatically change the number of locks that are used at run-time
depending on how much concurrency is detected.

Benchmark results comparing this option with the other ETS
optimization options are available here:

http://winsh.me/bench/ets_config_locks/ets_bench_result_lock_config.html