Make ETS hash use decentralized ctrs when w.c. is not true or false
This commit makes sure that the decentralized_counters ETS option is
turned on by default for all tables that has the write_concurrency
option set to something else than true or false. As before the
decentralized_counters option is still turned on by default when the
table type is set to ordered_set and the write concurrency option is
set to true.