compiler: Add a new instruction for creating binaries
Binary construction using the binary syntax currently uses multiple
instructions. That makes it hard to provide good information in the
exception if construction fails. It also makes it harder to optimize
construction in the JIT.

Therefore, introduce the instruction `bs_create_bin` that constructs
a binary in one go.

To be able to test the new instruction before implementing it in the
runtime system, force translation back to the old-style instructions
in the `beam_clean` pass.