docgen: Fix hidden docs for modules and functions
All modules and functions that do not have any documentation
should be marked as hidden. This is because we want to
EEP-48 doc chunks to be able to differentiate in between
modules/functions that are private and those that are just
not documented. In Erlang/OTP all modules/functions that are
public are documented, so we generate hidden docs for all
modules/functions that do not have any documentation.