Add GC check to find empty non-literal boxed terms on heap
This commit adds a check in debug builds of erts, every time a GC is
triggered. The check goes through terms that are on the GCed process'
heap. The emulator crashes, if the check finds a boxed pointer that
points to a non-literal (i.e., something that is not in a literal
area) that has nothing after the arity word is found. Currently, the
only type of boxed term that can have nothing after the arity word is
tuples of size 0 but this may change in the feature.

For testing convience a way to trigger the check prgrammatically is
also added to debug builds:

erts_debug:set_internal_state(available_internal_state, true),
erts_debug:get_internal_state(check_no_empty_boxed_non_literal_on_heap).