Restore save pointer on bad timeout value in receive
The message queue save pointer was not restored to the beginning of the
message queue when a bad timeout value was encountered in the receive. If
the timeout_value exception was caught, the following receive would not
be searched from the beginning of the message queue. This potentially
caused the process to hang forever even though a matching message was
present in the message queue.