erts: Optimize fun calls
By making local and external funs share the same structure, we can
remove a substantial amount of code from the `call_fun` family of
instructions.

As a result fun calls are now about 15% faster with the interpreter
on my machine, 10% faster with the x64 JIT on that same machine,
and 15% faster with the AArch64 JIT on my M1 Mac. The improvement
is a bit less pronounced for funs with free variables, and a bit
more for those without (including external funs).