Clean up code for raw RRs esp. regarding standard RRs
Enforce the idea that when you supply either numeric Class
or Type to inet_dns:encode/1, it is a raw record and the Data
*should* be bytes.

Also enforce that with symbolic Class and Type, the Data *should*
have correct format or else encoding will crash.  Previously
the code fell through to raw byte encoding if the format did
not match, which opens up for accidents.

Assume that the "standard RRs" that are the same for "all classes"
means for all known classes, not numeric (unknown) classes.

A numeric Class or Type is assumed to be raw, even if it should
be the same number as for a known symbolic one.