crypto: Make get_curve_cnt() less racy
As algo_curve_cnt and algo_curve_cnt_initialized are read
without any lock we are not guaranteed consistency. We may read
algo_curve_cnt_initialized as 1 and then algo_curve_cnt as 0.

Instead mark uninitialized with negative value and skip the separate
_initialized variable.

It's still not 100% kosher to read algo_curve_cnt without
synchronization but will work in practice.