erts_factory_undo: fix main heap not reset correctly bug
When initializing a heap factory with the function
erts_factory_proc_prealloc_init, the preallocated heap space can be
placed in a heap fragment if the main heap does not have enough
space. In that case, a pointer to the first allocated word in the
fragment is saved in the factory's heap_start field so that the heap
can be reset (if an erts_factory_undo call is made later) but no
pointer to the heap top of the main heap is saved. However, space on
the main heap can be allocated later by the factory if the function
erts_produce_heap is called with the factory as argument. Before this
fix there would then be no way to restore the main heap in an
erts_factory_undo call, because the previous main heap top would be
lost.

This commit fixes the issue described above by introducing a new field
in the factory struct that stores the top of the main heap at the time
the factory is initialized. This field is used to restore the main
heap top in the erts_factory_undo function.