sys_core_fold: Don't replace bad binary segments with error(badarg)
Keep warnings for binary construction that will fail, but don't
replace the construction with an `error(badarg)` call to allow the
runtime system to add extended information to the exception.

For example, `sys_core_fold` will no longer replace the following
code:

     <<F:7/float>>

with a call to `error(badarg)`.

However, note that v3_core still replaces obviously bad types and
sizes with a call to `error(badarg)`. For example, `v3_core` will
replace the following construction with a call to `error(badarg)`:

     <<atom:32>>