[snmp|agent|test] Tweaked mibs suite test
There is no test cases that would cause a node down, so such
an event is not our fault. Also, there is nothing we can do
about it. So if this happens => SKIP.
So, when tc_try/3 receives either 'nodedown' or an EXIT
with reason 'noconnection', skip the test case (instead of
failing).