erts: Print stackdump in crashdump/procdump with y prefix
In OTP-23 the placement of the top frame was changed and that
caused the stackdump printing algorithm to not prefix variables
in the first frame with y. i.e.

    0x00007f72440b76f0:N
    0x00007f72440b76f8:SReturn addr 0x44381898 (proc_lib:init_p_do_apply/3 + 72)
    y0:P<0.9.0>
    y1:P<0.10.0>
    0x00007f72440b7718:SReturn addr 0x66325748 (<terminate process normally>)

was written instead of

    y0:N
    0x00007f72440b76f8:SReturn addr 0x44381898 (proc_lib:init_p_do_apply/3 + 72)
    y0:P<0.9.0>
    y1:P<0.10.0>
    0x00007f72440b7718:SReturn addr 0x66325748 (<terminate process normally>)

The same issue was also fixed for erlang:process_display,
erlang:process_info(backtrace) and the ctrl-c proc dump.