docgen: Add prefix attribute for <datatype> name to include module
Ideally we would like to be able to write `-type erlang:type/()` in the
code, but that change would then need to be implemented in many many
places. So for now we just do this hack in the html docs to be
able to diffirentiate inbetween predifined types and user defined types.