erts: Fix testcase in previous commit
The test could not be loaded on 32-bit machines because the line
number wouldn't fit a signed word, so we'll lower it to a point
where it can be loaded.

We'll settle for this instead of fixing loading on 32-bit since
trying to load files with absurd line numbers is extremely
contrived.