jit: Fix rem guard bif segfault
In x86 rax and RET are the same values, so for rem guards the
value in RET was overwitten with the content of TMP[1] before
the test was done. This caused the rem to not fail as it should
and instead return the contents of TMP[1] which could be anything.