Make Erlang installation file system location independent
Previously shell scripts (e.g., `erl` and `start`) and the RELEASES file
for an Erlang installation depended on hard coded absolute path to the
installation's root directory. This made it cumbersome to move a
release to a different directory and can be problematic for platforms
such as Android (https://github.com/erlang/otp/pull/2879) where the
installation directory is unknown at compile time. This is fixed by:

* Changing the shell scripts so that they can dynamically find the
  `ROOTDIR`. The dynamically found `ROOTDIR` is selected if it differs
  from the hard-coded `ROOTDIR` and seems to point to a valid Erlang
  installation. The `dyn_erl` program has been changed so that it can
  return its absolute canonicalized path when given the `--realpath`
  argument (dyn_erl gets its absolute canonicalized path from the
  `realpath` POSIX function). The `dyn_erl`'s `--realpath`
  functionality is used by the scripts to get the root dir dynamically.

* Changing the `release_handler` module that reads and writes to the
  `RELEASES` file so that it prepends `code:root_dir()` whenever it
  encounters relative paths. This is necessary since the current
  working directory can be changed so it is something different than
  `code:root_dir()`.