Support a definition with non-zero arity in letrecs
Teach the `v3_kernel` and `beam_kernel_to_ssa` passes to handle
a `letrec` with the `letrec_goto` option, where the defined function
has an arity of one or higher. Also update Dialyzer to handle such
letrecs.

This enhanced `letrec` will come in handy for implementing
EEP 49 in the next commit.