Implement the compiler part of EEP 49
This commit implements the compiler part of EEP 49 (Value-Based Error
Handling Mechanisms).

There is one difference compared to the current version of the EEP 49
document, namely that the matching operator now is `?=` instead of
`<~`. Thus the rewritten Mnesia code from the EEP will now look like:

    commit_write(OpaqueData) ->
        maybe
            ok ?= disk_log:sync(OpaqueData#backup.file_desc),
            ok ?= disk_log:close(OpaqueData#backup.file_desc),
            ok ?= file:rename(OpaqueData#backup.tmp_file, OpaqueData#backup.file),
            {ok, OpaqueData#backup.file}
        end.

The reason for the change is that `<-` and `<~` look too similar to
each other (especially with some fonts and text sizes) and could be
easily confused. `?=` does not have that problem, and its similarity
to `=` suggests that it is a matching operator.

The preprocessor uses `?` to signal the start of a macro invocation.
Although `=` is a valid macro name, that does not cause any ambiguity,
because macro names that are not atoms or variables must be
single-quoted. Therefore, `?=` is not a valid invocation of the `=`
macro; the invocation of the `=` macro must be written like `?'='`.

The `maybe` ... `else` ... `end` construct is only allowed to be used
if the compiler option `{enable_feature,maybe}` has been given. If
not given, `maybe` and `else` will be interpreted as atoms, and code
that use those atoms can be compiled. As currently implemented, the
`enable_feature` option will not be recognized if given in the source
code for the module. This limitation will be lifted later (there is
ongoing work on a more robust and general implementation of
`enable_feature`).