Add a poor man's implementation of `{enable_feature,Feature}`
There is ongoing work, which will result in an EEP and a pull request,
for a way to conditionally enable new or experimental features in
the compiler and runtime system.

This commit implements this feature in a crude way, to allow us to
selectively make `maybe` and `else` keywords instead of atoms. That
will allow us to compile source code that use `maybe` and `else`
as atoms. (They are quite a lot of code in OTP that uses `maybe` as
an atom.)

Note that as implemented in this commit, for the
`{enable_feature,maybe}` to have any effect, it must be given as an
option to the compiler, **not** in the source code for the module.