Limit timeout value in calls to select()
We previously did not limit the timeout value when calling select()
which could cause select() to fail when passed a huge timeout value
causing a runtime system crash.

We currently only use select() on Darwin which accepts a timeout value
of up to 100 million seconds. However, according to posix, select()
implementations are only required to provide 31 days as max timeout
value. We therefore limit the timeout value to 31 days.