stdlib: Add helper to support round tripping using erl_pp:form/{1,2}
The Erlang compiler will, when expanding records to tuples, introduce
new variables in the abstract representation. As the expansion is done
on the abstract representation, the compiler can safely name the new
variables with names that are not syntactically valid in Erlang source
code (the name starts with a lowercase letter), thus ensuring the
uniqueness of the new names.

The above strategy leads to problems if a user uses the compiler to
produce the abstract representation (using options `E` and `binary`),
transforms it and intends to use the functions in erl_pp to produce
source code, as the generated variable names will be printed as atoms.

This patch introduces an utility function, `erl_pp:legalize_vars/1`,
which when given the abstract representation of a function, will
return an equivalent function where all variables will have
syntactically valid names. As outputting Erlang source code from the
abstract representation is a niche application, the solution with an
utility function, instead of modifying compiler passes, was chosen to
not slow down the common use case.