yecc: Avoid compilation errors if atoms become reserved words
In code generated by `yecc`, only atoms known to be reserved words
would be quoted. If a generated module is later compiled with a
version of the compiler that has made one of the used atoms a
reserved word, compilation would fail.

Avoid the problem by having `yecc` code quote all atoms in the
generated code.