erts: Fix integer overflow problem in WSTACK
In the testcase binary_SUITE:t2b_deterministic a very large
hash map is created and encoded. This lead to a WSTACK_RESERVE
call with an N value of 4 000 000 which made the pointer wrap around
zero and thus the stack was not grown as it should.

As far as I can tell this can only happen in the new
deterministic term_to_binary.