observer: Skip testing on sunos
With the current wxWidgets/gtk and/or test X-server version,
the tests crash when closing the window.

This only happens if the application sets an icon on the top-frame,
so something is broken in those libraries when deleting the window.

Skip tests for now, since both observer and crashdump_viewer sets an
icon.