Change default unspecified ssh line ending to CRLF
If an SSH client does not specify what line endings to use, use CRLF.
This fixes the built-in Windows ssh client that currently reports as
"OpenSSH_for_Windows_8.1p1, LibreSSL 3.0.2".

The reason for changing the default rather than adding another special
case is that OpenSSH defaults to sending CR characters when unspecified.
See the use of ONLCR at [1]. This mirrors that behavior.

Interestingly, I found ONLCR is normally on with my testing with OpenSSH
SSH to connect between MacOS and Linux, so this may not be an unusual
setting anyway for non-Windows computers. Verification was done by
capturing output and checking `stty -a | grep onlcr`.

Tested on Windows (Putty and commandline ssh in Powershell), Linux and
MacOS.

[1] https://github.com/openssh/openssh-portable/blob/2dc328023f60212cd29504fc05d849133ae47355/openbsd-compat/bsd-openpty.c#L194