Correct scope of variables
Consider this example:

    main(_) ->
        _ = (X = 42) + fun() -> X = 0 end(),
        io:format("~p\n", [X]).

When run as an `escript`, we will get different results depending
on whether we run it using `erl_eval` or using the Erlang compiler.

Here is the result using `erl_eval` (the default):

    $ escript t.erl
    42

Here is the result using the compiler:

    $ escript -c t.erl
    escript: exception error: no match of right hand side value 0

At least one of the results must be wrong. Which?

The compiler is wrong.

A binary operator (such as `+`) is allowed to evaluate its operands
in any order. Thus, code that depends on the left hand side of
a binary operator being evaluated before the right hand side is
illegal and will be rejected by the compiler. For example:

    bar(A) ->
        (X = A) + (2 * X).

Compilation will fail with the following message:

    t.erl:5:20: variable 'X' is unbound
    %    5|     (X = A) + (2 * X).
    %     |                    ^

Therefore, when we have the expression:

    (X = 42) + fun() -> X = 0 end()

the variable `X` is not bound when evaluating the right hand side
operand for `+` and the `X` bound in the fun body is another variable
named `X`, unrelated to the `X` bound on the left hand side of `+`.

This commit corrects the compiler so that it generates code that
evalutes the example in the same way as `erl_eval` does.

Closes #5379