Somewhat improve pretty printing of binaries in Dialyzer
When running Dialyzer on this example:

    -spec foo() -> ok.
    foo() ->
        case foo of
          <<"bar">> -> ok;
          _ -> ok
        end.

the following error would be generated:

    The pattern <<6447474:24>> can never match the type 'foo'

That message is very confusing. Unfortunately, Dialyzer has only
access to Core Erlang version of the code, so there is no easy way to
exactly pretty print code fragments. However, the optimization of
strings that produces those huge numbers can be disabled when Dialyzer
invokes the compiler, producing the following message:

    The pattern <<98,97,114>> can never match the type 'foo'

Closes #5429.