[snmp|manager|test] Use port 0 (zero) instead of 5000
In these tests there is actually no communication, so there is no
need to use a specific port. Therefor, make the system choose
(by specifying potr = 0).

The reason for this is that it seems that on some systems, the
port (5000) seems to be still "used" (from the previous tests).

OTP-17783