compiler: Preserve extended error information
The compiler currently replaces known failing calls with a call to
`error(badarg)`. Example:

    foobar() ->
        atom_to_binary(42).

The compiler rewrites that to:

    foobar() ->
        error(badarg).

With the introduction of extended error information (EEP 54), that
"optimization" results in no extended error information being shown:

    1> t:foobar().
    ** exception error: bad argument
         in function  t:foobar/0 (t.erl, line 5)

To ensure that extended error information is always shown, update
the compiler so that it will keep the known failing call (but still
emit a warning). Thus:

    2> t:foobar().
    ** exception error: bad argument
         in function  atom_to_binary/1
            called as atom_to_binary(42)
            *** argument 1: not an atom

Closes #5440.