yeccpre.hrl: Never crash when printing non-standard tokens
The standard prologue file (`yeccpre.hrl`) for yecc could crash while
attempting to report a syntax error involving a token not produced by
the `erl_scan` module.

For example, the token `{string,0,<<"hello">>}` looks like a string
token produced by `erl_scan`, but the string value is a binary instead
of the expected list and the default implementation of
`yecctoken_to_string/1` would crash when asked to print it.

Make `yecctoken_to_string/1` more resilient against crashing by
printing the entire token as is if there is a crash when printing
it in the usual way.

(Note: A parser that uses its own non-standard scanner should use its
own customized prologue file if nicely printed tokens are desired.)