erts: Try to make dump_SUITE:abort_signal more stable
We do this by using internal state sleeping instead of
spinning.

The idea is that it will be more likely for a process
to be scheduled in when the abort comes if they sleep
in internal state.

An effect of this was that the system became much less
responsive, so we make it so that the number of dist
messages are limited and also bind all actions done to
seperate schedulers.