compiler: Improve handling of container types
We used to determine the type of `get_tuple_element` at the time
of extraction, which is simple but sometimes throws away type
information when used on tuple unions.

This normally doesn't cause any issues other than slightly less
optimized code, but would crash the type pass in rare cases.
Consider the following SSA:

```
 %% (Assume _0 is either `{a, 1}`, {b, 2}, or `{c, {d}}`)
 0:
     _1 = get_tuple_element _0, `0`
     _2 = get_tuple_element _0, `1`
     switch _1, ^3, [
         { `a`, ^1 },
         { `b`, ^2 }
     ]
 1: %% ... snip ...
 2: %% ... snip ...
 3:
     _3 = get_tuple_element _0, `1`
     @ssa_bool = is_tagged_tuple _3, 1, `d`
     br @ssa_bool ^4, ^1234
 4:
     _4 = get_tuple_element _3, `0`
```

In block 0 we determine that the type of _1 is `a | b | c` and
that _2 is `1 | 2 | {d}`. From this, we know that _0 is `{a, 1}`
in block 1, `{b, 2}` in block 2, and `{c, {d}}` in block 3.

In block 3, we remove the `is_tagged_tuple` test since it's
redundant.

... but then the compiler replaces _3 with _2 since they're the
same value. However, the type is still the same old `1 | 2 | {d}`
we got in block 0, which is not safe for `get_tuple_element`,
crashing the type pass.

This commit fixes this by determining the type of
`get_tuple_element` when the result is used instead of when it was
extracted, giving the correct type `{d}` in block 4.