Remove undocumented {value, Value, Ann} in erl_eval
The shell relies on an implicit contract
with erl_eval where value triplet is allowed.
Unfortunately, to support this, evaluating
anonymous functions would require a look-ahead
traversal to hide any value triplet from
erl_lint and this affected performance negatively.

This patch removes this undocumented contract
and removes the lookahead. To address this,
we took different approaches:

  1. v(N) was expanded to a value before evaluated.
     Now, we process and validate the arguments,
     but communicate with the shell process to get
     the value, similar to how history() works

  2. References, Ports, and PIDs were converted to
     variables and then replaced to values. Now,
     we expand `#PID<0.13.0>` into an expression
     such as `erlang:list_to_pid("#PID<0.13.0>").`

  3. Finally, the prompt evaluation would use values,
     but it only had to pass `[{history,integer()}]`,
     which can be manually converted to AST

With this change, evaluating anonymous functions in
a loop gets considerably faster, up to 10% in a
macro benchmark.