erts: Fix spurious exceptions in 'fnegate'
If we were to enter an 'fnegate' instruction with a non-finite
value in xmm0 by accident, we would erroneously throw an
exception.

I'm not certain whether this can occur in practice: 'fnegate' is
always preceded by another floating-point instruction that either
leaves a valid value in xmm0 or throws an exception (`fconv`,
`fload`, `fadd`, etc).

I can't think of a valid instruction sequence that would cause it,
but we'll fix it anyway just in case.