jit: Optimize yield tests
We no longer use `CALL` to figure out the yield address, so we may
as well get rid of it. While we're at it, we'll shrink the
instruction by branching directly into the shared fragment.

We'll use the same method on AArch64 as well as it's a bit easier
to follow.