observer/etop: Calculate field width with available columns
The field widths are calculated depending on what the underlying io
driver reports as max column width.

The fields are allowed to use the column width they need if they column
budget allows it. The field widths are deducted in priority order and if
there is space left the calculated width is used otherwise the default
field width is used.

If the underlying io driver can't give any column width, the default
column width budget of 89 characters is used. This is the legacy hard
coded width, plus two spaces between reductions, memory, and msgq
fields.

Fixes #5080

Signed-off-by: Per Andersson <avtobiff@foo.nu>