ct: Optimize surefire cth to keep state in seperate process
When running a suite with many processes the cth state
could become very large. Since the state is copied to the
testcase process each time a cth is called that because
very expensive. So instead we create a seperate process
that keeps the state of the surefire cth so that it does
not have to be copied.