Fix timing bug in ensure_engine_loaded
When two ensure_engine_loaded() calls were done in parallell there
was a possibility that a crypto lib function was called by both instead
of just one of them.

This is solved by moving the implementation from erlang down into a nif
function that uses a mutex to protect the sensitive part.