public_key,ssl: add password fun for decoding keyfiles
Implementing password fun option for ssl connect API

This change allows password option to accept as a value
either a string or a fun. Added specific tests and
documentation.

The motivation for this change is to better protect a
private key. If the private key is protected by password the
password may be retreived using user supplied function therefore
enabling storing password in a secure vault or something similar.