beam_call_types: Improve range optimizations
Add ranges to the type for functions that returns a size for an
Erlang term. Since any term is limited by the size of memory, it
is possible to define an upper limit that can never be reached in
practice in the foreseeable future.

Also improve propagation of ranges through some more operations.