erts: Fix spawn_request trace bug
Since the args of spawn_request may end up in a trace
we need to allocate them on the heap of the spawning process.
We cannot use a temporary heap for this as
allocates on normal heap and this will lead to stray pointers
on the heap during gc. Stray pointers are not always a problem
as any dead data is ignored, but if the pointers and up in a
heap fragments they are not ignored...