update_deprecations: The script didn't handle '-deprecated([f/1])'
which caused the atom 'undefined' to be the Text in update_deprecations:make_xml_info/2.
According to /home/uabhnil/Projects/OTP/otp1/HOWTO/DEPRECATE.md that 'undefined' should be changed to
"see the documentation for details." by that script.
The internal wiki page has the same opinion as the script.