beam_trim: Reduce compile time for huge straight-line functions
For a huge function with straight-line code, the `beam_trim` pass
could take forever, mainly because of the ineffcient way of
calculating the use information in the `frame_layout()` function.

This commit reduces the time for the `beam_trim` of the sample code
in #5140 to about 1 second on my computer (down from at least
5 minutes).

While at it, also add support for new instructions in OTP 25, and
fix a minor bug resulting in some trimming opportunities were being
ignored.

Co-authored-by: John Högberg <john@erlang.org>