compiler: Reflect all exceptions in SSA control flow
We cheated a bit on tail calls outside of try/catch expressions,
reasoning that we've left the current function by the time an
exception is thrown, so there's no need to represent it in the
control flow.

This is logically consistent, but makes it difficult to figure out
whether a function throws.

This commit fixes that by having tail calls follow the
`{succeeded,body}` convention used by all other instructions that
may throw exceptions.