[snmp|agent|test] Add a "sanity check" for framework test case
On some of our slow (and/or "unstable) machines, time checks may
produce results that are untrustworthy.
So, as a sort of sanity check, we check how long time it took
to get the snmpEngineTime. This should normally only take a couple
of milli sec (at most 100ms). So, if it takes more tha a second
then we can no longer trust the results. Therefor SKIP instead.

Also update copyright end date.