Fix coercion of integer literals to floats in binary matching
In both binary construction and matching, literal integer values
are allowed in float segments, and they are automatically coerced
to floats.

Thus, the compiler automatically rewrites code such as the following:

    foo() ->
        <<0/float-native>> = <<0/float-native>>.

to:

    foo() ->
        <<0.0/float-native>> = <<0.0/float-native>>.

However, if the order of `native` and `float` were switched like this:

    bar() ->
        <<0/native-float>> = <<0/native-float>>.

the compiler would fail to coerce the `0` in the binary pattern and would
end up with:

    bar() ->
        <<0/native-float>> = <<0.0/native-float>>.

which would obviously not match.